+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [1.1.0] - 2020-07-17
+
+### Added
+
+- Script now populates 'Tags' column in spreadsheet containing a list of values for each event.
+- Config sheet has new property 'Ignore Tags'. Value is a comma-seperated list of tags. Any event that has a tag matching any of the ignored tags will be excluded from the update process.
+  - Default is `Xbox, PS4`, filtering out console exclusives.
+
 ## [1.0.0] - 2020-07-12
 
 ### Added